Address PR review comments
- Move computeCardStatusFlags and isItemPRMerged to
  src/shared/utils/trackerCardStatus.ts. Screens and dialogs are sibling
  layers — having a dialog import upward into a screen module was a
  layering inversion. Both callers now import from the shared utility.
- Fix consent-gate vs stale-ready precedence: when live aiStatus ===
  'waiting', readyToAdvance must also be suppressed, otherwise the green
  "Ready ✓" branch in getTrackerCardDisplayState wins over the yellow
  consent-gate render. Adds the matching test case.
